requirements.txt
fastapi~=0.115.4
pydantic~=2.9.2
starlette~=0.41.2
loguru~=0.7.2
SQLAlchemy~=2.0.36
uvicorn~=0.32.0
asyncpg~=0.30.0
passlib~=1.7.4
bcrypt~=4.2.0
PyJWT~=2.9.0
python-dotenv~=1.0.1